Rename Marine invisibility to generic Entity
