+++ Content.Client/_RMC14/Marines/Invisibility/MarineInvisibilityVisualsSystem.cs
@@ -1,4 +1,4 @@
-﻿using Content.Shared._RMC14.Marines.Invisibility;
+﻿using Content.Shared._RMC14.Stealth;
 using Robust.Client.GameObjects;
 
 namespace Content.Client._RMC14.Marines.Invisibility;
@@ -7,10 +7,10 @@ public sealed class MarineInvisibilityVisualsSystem : EntitySystem
 {
     public override void Update(float frameTime)
     {
-        var invisible = EntityQueryEnumerator<MarineTurnInvisibleComponent, SpriteComponent>();
+        var invisible = EntityQueryEnumerator<EntityTurnInvisibleComponent, SpriteComponent>();
         while (invisible.MoveNext(out var uid, out var comp, out var sprite))
         {
-            var opacity =  TryComp<MarineActiveInvisibleComponent>(uid, out var activeInvisible) ? activeInvisible.Opacity : 1;
+            var opacity =  TryComp<EntityActiveInvisibleComponent>(uid, out var activeInvisible) ? activeInvisible.Opacity : 1;
             sprite.Color = Color.Transparent.WithAlpha(opacity);
         }
     }
