+++ Content.Shared/_RMC14/Stealth/EntityActiveInvisibleComponent.cs
@@ -1,9 +1,9 @@
 ﻿using Robust.Shared.GameStates;
 
-namespace Content.Shared._RMC14.Marines.Invisibility;
+namespace Content.Shared._RMC14.Stealth;
 
 [RegisterComponent, NetworkedComponent, AutoGenerateComponentState, AutoGenerateComponentPause]
-public sealed partial class MarineActiveInvisibleComponent : Component
+public sealed partial class EntityActiveInvisibleComponent : Component
 {
     [DataField, AutoNetworkedField]
     public float Opacity = 0.1f;
